Upgraded blocksize-capital to v3 framework
