+++ packages/core/legos/package.json
@@ -52,7 +52,6 @@
     "@chainlink/blockchair-adapter": "workspace:*",
     "@chainlink/blockcypher-adapter": "workspace:*",
     "@chainlink/blocksize-capital-adapter": "workspace:*",
-    "@chainlink/blocksize-capital-test-adapter": "workspace:*",
     "@chainlink/blockstream-adapter": "workspace:*",
     "@chainlink/bob-adapter": "workspace:*",
     "@chainlink/bravenewcoin-adapter": "workspace:*",
