+++ packages/core/legos/src/sources.ts
@@ -21,7 +21,6 @@ import * as blockchain_com from '@chainlink/blockchain.com-adapter'
 import * as blockchair from '@chainlink/blockchair-adapter'
 import * as blockcypher from '@chainlink/blockcypher-adapter'
 import * as blocksize_capital from '@chainlink/blocksize-capital-adapter'
-import * as blocksize_capital_test from '@chainlink/blocksize-capital-test-adapter'
 import * as blockstream from '@chainlink/blockstream-adapter'
 import * as bob from '@chainlink/bob-adapter'
 import * as bravenewcoin from '@chainlink/bravenewcoin-adapter'
@@ -188,7 +187,6 @@ export default {
   blockchair,
   blockcypher,
   blocksize_capital,
-  blocksize_capital_test,
   blockstream,
   bob,
   bravenewcoin,
