+++ packages/core/legos/tsconfig.json
@@ -80,9 +80,6 @@
       "path": "../../sources/blocksize-capital"
     },
     {
-      "path": "../../sources/blocksize-capital-test"
-    },
-    {
       "path": "../../sources/blockstream"
     },
     {
