--- packages/sources/blocksize-capital-test/CHANGELOG.md
@@ -1,67 +1,0 @@
-# @chainlink/blocksize-capital-test-adapter
-
-## 1.1.8
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.1.7
-
-### Patch Changes
-
-- 460427afe: Fixed ws response parsing
-
-## 1.1.6
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.1.5
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.1.4
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.1.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.1.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.1.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.0.1
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
-
-## 1.0.0
-
-### Major Changes
-
-- fb94c36d6: Add blocksizecapital v3 EA
