--- packages/sources/blocksize-capital-test/src/config/index.ts
@@ -1,18 +1,0 @@
-import { AdapterConfig } from '@chainlink/external-adapter-framework/config'
-
-export const defaultEndpoint = 'price'
-
-export const config = new AdapterConfig({
-  API_KEY: {
-    description: 'The Blocksize Capital API key',
-    type: 'string',
-    required: true,
-    sensitive: true,
-  },
-  WS_API_ENDPOINT: {
-    description: 'The default WebSocket API base url',
-    type: 'string',
-    required: false,
-    default: 'wss://data.blocksize.capital/marketdata/v1/ws',
-  },
-})
