--- packages/sources/blocksize-capital/schemas/env.json
@@ -1,18 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/blocksize-capital-adapter.json",
-  "title": "Blocksize Capital Adapter",
-  "description": "[Blocksize Capital API](https://data-dev.blocksize.capital/marketdata/docs/#blocksize-marketdata-api-v1)",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "API key to use"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
