--- packages/sources/blocksize-capital/src/config/includes.json
@@ -1,1 +1,0 @@
-{}
