--- packages/sources/blocksize-capital/src/config/symbols.json
@@ -1,1 +1,0 @@
-{}
