+++ packages/sources/blocksize-capital/test/integration/adapter.test.ts
@@ -1,77 +1,81 @@
-import { AdapterRequest, FastifyInstance } from '@chainlink/ea-bootstrap'
-import request, { SuperTest, Test } from 'supertest'
-import { server as startServer } from '../../src'
-import { mockLoginResponse, mockSubscribeResponse, mockUnsubscribeResponse } from './fixtures'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
+import { AdapterRequestBody, sleep } from '@chainlink/external-adapter-framework/util'
+import { Server } from 'mock-socket'
 import { AddressInfo } from 'net'
-import {
-  mockWebSocketProvider,
-  mockWebSocketServer,
-  MockWsServer,
-  mockWebSocketFlow,
-  setEnvVariables,
-} from '@chainlink/ea-test-helpers'
-import { WebSocketClassProvider } from '@chainlink/ea-bootstrap/dist/lib/middleware/ws/recorder'
-import { DEFAULT_BASE_WS_URL } from '../../src/config'
-
-let oldEnv: NodeJS.ProcessEnv
+import request, { SuperTest, Test } from 'supertest'
+import { createAdapter, mockWebSocketProvider, mockWebSocketServer, setEnvVariables } from './setup'
 
 describe('websocket', () => {
-  let mockedWsServer: InstanceType<typeof MockWsServer>
-  let fastify: FastifyInstance
-  let req: SuperTest<Test>
-
   beforeAll(async () => {
-    oldEnv = JSON.parse(JSON.stringify(process.env))
-    process.env.WS_ENABLED = 'true'
-
-    if (!process.env.RECORD) {
-      mockedWsServer = mockWebSocketServer(DEFAULT_BASE_WS_URL)
-      mockWebSocketProvider(WebSocketClassProvider)
-      process.env.API_KEY = 'fake-api-key'
-      process.env.WS_SUBSCRIPTION_TTL = '100'
-    } else {
-      process.env.WS_SUBSCRIPTION_TTL = '3000'
-    }
-
-    fastify = await startServer()
-    req = request(`localhost:${(fastify.server.address() as AddressInfo).port}`)
+    const mockDate = new Date('2022-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
   })
 
   afterAll((done) => {
+    spy.mockRestore()
+    done()
+  })
+
+  let fastify: ServerInstance | undefined
+  let req: SuperTest<Test>
+  let spy: jest.SpyInstance
+  let mockWsServer: Server | undefined
+  const wsEndpoint = 'wss://data.blocksize.capital/marketdata/v1/ws'
+
+  jest.setTimeout(10000)
+
+  const data: AdapterRequestBody = {
+    data: {
+      base: 'ETH',
+      quote: 'EUR',
+    },
+  }
+
+  let oldEnv: NodeJS.ProcessEnv
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env['WS_SUBSCRIPTION_TTL'] = '1000'
+    process.env['CACHE_MAX_AGE'] = '1000'
+    process.env['CACHE_POLLING_MAX_RETRIES'] = '0'
+    process.env['METRICS_ENABLED'] = 'false'
+    process.env['WS_ENABLED'] = 'true'
+    process.env['WS_API_ENDPOINT'] = wsEndpoint
+    process.env['API_KEY'] = 'fake-api-key'
+
+    const mockDate = new Date('2022-05-10T16:09:27.193Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+
+    // Start mock web socket server
+    mockWebSocketProvider(WebSocketClassProvider)
+    mockWsServer = mockWebSocketServer(wsEndpoint)
+
+    fastify = await expose(createAdapter())
+    req = request(`http://localhost:${(fastify?.server.address() as AddressInfo).port}`)
+
+    // Send initial request to start background execute
+    await req.post('/').send(data)
+    await sleep(5000)
+  })
+
+  afterAll((done) => {
+    spy.mockRestore()
     setEnvVariables(oldEnv)
-    fastify.close(done)
+    mockWsServer?.close()
+    fastify?.close(done())
   })
 
   describe('price endpoint', () => {
     it('should return success', async () => {
-      const data: AdapterRequest = {
-        id: '1',
-        data: {
-          base: 'ETH',
-          quote: 'EUR',
-        },
-      }
-
-      let flowFulfilled = Promise.resolve(true)
-      if (!process.env.RECORD) {
-        flowFulfilled = mockWebSocketFlow(mockedWsServer, [
-          mockLoginResponse,
-          mockSubscribeResponse,
-          mockUnsubscribeResponse,
-        ])
-      }
-
-      const response = await req
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-
+      const makeRequest = () =>
+        req
+          .post('/')
+          .send(data)
+          .set('Accept', '*/*')
+          .set('Content-Type', 'application/json')
+          .expect('Content-Type', /json/)
+      const response = await makeRequest()
       expect(response.body).toMatchSnapshot()
-
-      await flowFulfilled
-    })
+    }, 30000)
   })
 })
