+++ packages/sources/blocksize-capital/tsconfig.json
@@ -5,10 +5,5 @@
     "rootDir": "src"
   },
   "include": ["src/**/*", "src/**/*.json"],
-  "exclude": ["dist", "**/*.spec.ts", "**/*.test.ts"],
-  "references": [
-    { "path": "../../core/test-helpers" },
-    { "path": "../../core/bootstrap" },
-    { "path": "../../core/test-helpers" }
-  ]
+  "exclude": ["dist", "**/*.spec.ts", "**/*.test.ts"]
 }
