+++ packages/tsconfig.json
@@ -192,9 +192,6 @@
       "path": "./sources/blocksize-capital"
     },
     {
-      "path": "./sources/blocksize-capital-test"
-    },
-    {
       "path": "./sources/blockstream"
     },
     {
